Code Restructuring + add extra tests
