+++ core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/ComposablePreviewScanner.kt
@@ -1,9 +1,10 @@
 package sergio.sastre.composable.preview.scanner.core.scanner
 
 import io.github.classgraph.ClassGraph
-import io.github.classgraph.ScanResult
-import sergio.sastre.composable.preview.scanner.core.scanner.classloader.classpath.Classpath
-import sergio.sastre.composable.preview.scanner.core.scanner.classloader.classpath.previewfinder.ClasspathPreviewsFinder
+import sergio.sastre.composable.preview.scanner.core.scanner.config.ClassGraphSourceScanner
+import sergio.sastre.composable.preview.scanner.core.scanner.config.classloader.classpath.Classpath
+import sergio.sastre.composable.preview.scanner.core.scanner.config.classloader.classpath.previewfinder.ClasspathPreviewsFinder
+import sergio.sastre.composable.preview.scanner.core.scanner.config.SourceScanner
 import sergio.sastre.composable.preview.scanner.core.scanresult.RequiresLargeHeap
 import sergio.sastre.composable.preview.scanner.core.scanresult.filter.ScanResultFilter
 import java.io.File
@@ -11,78 +12,45 @@ import java.io.InputStream
 
 /**
  * Core Component to scan for Previews
- * @param defaultPackageTreesOfCrossModuleCustomPreviews // TODO
+ * @param defaultPackageTreesOfCrossModuleCustomPreviews package where external previews
+ * (i.e. previews defined in another dependency or module) can be found like those in "androidx.compose.ui.tooling.preview"
  */
 abstract class ComposablePreviewScanner<T>(
     private val findComposableWithPreviewsInClass: ClasspathPreviewsFinder<T>,
     private val defaultPackageTreesOfCrossModuleCustomPreviews: List<String> = emptyList()
-) {
+) : SourceScanner<T> {
 
-    private var updatedClassGraph = ClassGraph()
-        .ignoreMethodVisibility()
-        .enableClassInfo()
-        .enableMethodInfo()
-        .enableAnnotationInfo()
-        .enableMemoryMapping()
+    private var updatedClassGraph =
+        ClassGraph()
+            .ignoreMethodVisibility()
+            .enableClassInfo()
+            .enableMethodInfo()
+            .enableAnnotationInfo()
+            .enableMemoryMapping()
 
-    /**
-     * Scan previews in all packages, including those of external dependencies.
-     */
-    @RequiresLargeHeap
-    fun scanAllPackages(): ScanResultFilter<T> {
-        return ScanResultFilter(
-            updatedClassGraph.scan(),
-            findComposableWithPreviewsInClass,
-        )
-    }
+    private val classGraphSourceScanner =
+        ClassGraphSourceScanner(updatedClassGraph, findComposableWithPreviewsInClass)
 
     /**
-     * Scan previews in the given packageTrees
+     * Prepares the scanner to find previews from a Source Set like 'screenshotTest', 'androidTest', 'main' or a custom one via the given sourceSetClasspath
+     * It uses compiled classes of that source set.
+     * Check SourceSetClasspath to find their locations under the /build folder of the corresponding module.
      *
-     * @param packageTrees where we want to scan previews
-     */
-    fun scanPackageTrees(vararg packageTrees: String): ScanResultFilter<T> {
-        if (packageTrees.isEmpty()) {
-            throw IllegalArgumentException("packages must not be empty. For that, use scanAllPackages() instead")
-        }
-        updatedClassGraph = updatedClassGraph.acceptPackages(*packageTrees)
-        return ScanResultFilter(
-            updatedClassGraph.scan(),
-            findComposableWithPreviewsInClass,
-        )
-    }
-
-    /**
-     * Scan previews in the given ‘include‘ packageTrees, excluding the 'exclude' packageTrees
+     * Make sure those compiled classes exist and are up to date before scanning them.
+     * For that you can execute ./gradlew :<module>:compile<variant><sourceSet>Kotlin,
+     * for instance: ./gradlew :mymodule:compileReleaseScreenshotTestKotlin
      *
-     * @param include where we want to scan previews
-     * @param exclude where we do not want to scan previews, even though they were inside the included packageTrees
-     */
-    fun scanPackageTrees(include: List<String>, exclude: List<String>): ScanResultFilter<T> {
-        if (include.isEmpty()) {
-            throw IllegalArgumentException("include packages must not be empty. For that, use scanAllPackages() instead")
-        }
-        updatedClassGraph = updatedClassGraph
-            .acceptPackages(*include.toTypedArray())
-            .rejectPackages(*exclude.toTypedArray())
-        return ScanResultFilter(
-            updatedClassGraph.scan(),
-            findComposableWithPreviewsInClass,
-        )
-    }
-
-    /**
-     * Scan previews from a sourceSet like 'screenshotTest' or 'androidTest', 'main' or a custom one via the given sourceSetClasspath
-     *
-     * @param sourceSetClasspath the classpath pointing to the package where compiled classes of a sourceSet are located
-     * @param packageTreesOfCrossModuleCustomPreviews Package where you've defined custom Preview annotations that are present in this module as TRANSITIVE dependency.
-     * As general rule, this can be left empty. Only use it if y
+     * @param sourceSetClasspath the classpath pointing to the package where compiled classes of a Source Set are located
+     * @param packageTreesOfCrossModuleCustomPreviews package where external previews (i.e. previews defined in another dependency or module)
+     * different can be found. Previews under "androidx.compose.ui.tooling.preview", like @PreviewLightDark, do not need to be added here.
+     * In most projects, you can leave this empty unless you see some Previews missing
      */
     fun setTargetSourceSet(
         sourceSetClasspath: Classpath,
         packageTreesOfCrossModuleCustomPreviews: List<String> = emptyList()
-    ) = apply {
-        val absolutePath = File(sourceSetClasspath.buildDir, sourceSetClasspath.packagePath).absolutePath
+    ): ClassGraphSourceScanner<T> {
+        val absolutePath =
+            File(sourceSetClasspath.rootDir, sourceSetClasspath.packagePath).absolutePath
         findComposableWithPreviewsInClass
             .applyOverridenClasspath(sourceSetClasspath.packagePath)
             .applyCrossModuleCustomPreviewPackageTrees(
@@ -90,39 +58,61 @@ abstract class ComposablePreviewScanner<T>(
             )
 
         updatedClassGraph.overrideClasspath(absolutePath)
+
+        return ClassGraphSourceScanner(
+            classGraph = updatedClassGraph,
+            findComposableWithPreviewsInClass = findComposableWithPreviewsInClass,
+        )
     }
+
+    /**
+     * Scan previews in all packages, including those of external dependencies.
+     */
+    @RequiresLargeHeap
+    override fun scanAllPackages(): ScanResultFilter<T> =
+        classGraphSourceScanner.scanAllPackages()
+
+    /**
+     * Scan previews in the given packageTrees
+     *
+     * @param packageTrees where we want to scan previews
+     */
+    override fun scanPackageTrees(vararg packageTrees: String): ScanResultFilter<T> =
+        classGraphSourceScanner.scanPackageTrees(*packageTrees)
+
+    /**
+     * Scan previews in the given ‘include‘ packageTrees, excluding the 'exclude' packageTrees
+     *
+     * @param include where we want to scan previews
+     * @param exclude where we do not want to scan previews, even though they were inside the included packageTrees
+     */
+    override fun scanPackageTrees(
+        include: List<String>,
+        exclude: List<String>
+    ): ScanResultFilter<T> =
+        classGraphSourceScanner.scanPackageTrees(include, exclude)
 
     /**
      * Scan previews in the given file
      *
      * @param jsonFile a json file that was generated by using ScanResultDump.dumpScanResultToFile(fileName)
      */
-    // TODO -> Allow to dump also customPreviews
-    fun scanFile(jsonFile: File): ScanResultFilter<T> {
-        val scanResult = ScanResult.fromJSON(jsonFile.bufferedReader().use { it.readLine() })
-        return ScanResultFilter(
-            scanResult,
-            findComposableWithPreviewsInClass,
-        )
-    }
+    override fun scanFile(jsonFile: File): ScanResultFilter<T> =
+        classGraphSourceScanner.scanFile(jsonFile)
 
     /**
      * Scan for previews in the given InputStream.
      *
-     * @param targetInputStream to support accessible by instrumentation tests, like getInstrumentation().context.assets.open(fileName).
+     * @param targetInputStream to make them accessible in instrumentation tests, like getInstrumentation().context.assets.open(fileName).
      * Such file was generated previously by using ScanResultDump.dumpScanResultToFileInAssets(fileName)
      *
+     * @param customPreviewsInfoInputStream to make custom previews defined in a dependency or module,
+     * like those in "androidx.compose.ui.tooling.preview", available in instrumentation tests.
+     * Such file was generated previously by using ScanResultDump.dumpScanResultToFileInAssets(customPreviewsPackageTrees, customPreviewsFileName)
      */
-    fun scanFile(
+    override fun scanFile(
         targetInputStream: InputStream,
         customPreviewsInfoInputStream: InputStream
-    ): ScanResultFilter<T> {
-        val scanResult = ScanResult.fromJSON(targetInputStream.bufferedReader().use { it.readLine() })
-        findComposableWithPreviewsInClass
-            .applyCustomPreviewsScanResult(ScanResult.fromJSON(customPreviewsInfoInputStream.bufferedReader().use { it.readLine() }))
-        return ScanResultFilter(
-            scanResult,
-            findComposableWithPreviewsInClass,
-        )
-    }
+    ): ScanResultFilter<T> =
+        classGraphSourceScanner.scanFile(targetInputStream, customPreviewsInfoInputStream)
 }
