--- core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/classloader/ClassLoader.kt
@@ -1,7 +1,0 @@
-package sergio.sastre.composable.preview.scanner.core.scanner.classloader
-
-import io.github.classgraph.ClassInfo
-
-internal interface ClassLoader {
-    fun loadClass(classInfo: ClassInfo): Class<*>
-}
