--- core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/classloader/ReflectionClassLoader.kt
@@ -1,8 +1,0 @@
-package sergio.sastre.composable.preview.scanner.core.scanner.classloader
-
-import io.github.classgraph.ClassInfo
-
-internal class ReflectionClassLoader: ClassLoader {
-    override fun loadClass(classInfo: ClassInfo): Class<*> =
-        Class.forName(classInfo.name, false, this.javaClass.classLoader)
-}
