--- core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/classloader/classpath/Classpath.kt
@@ -1,6 +1,0 @@
-package sergio.sastre.composable.preview.scanner.core.scanner.classloader.classpath
-
-data class Classpath(
-    val packagePath: String,
-    val buildDir: String = "${System.getProperty("user.dir")}/build"
-)
