--- core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/classloader/classpath/previewfinder/overriden/annotationloader/CustomPreviewAnnotationLoader.kt
@@ -1,8 +1,0 @@
-package sergio.sastre.composable.preview.scanner.core.scanner.classloader.classpath.previewfinder.overriden.annotationloader
-
-import io.github.classgraph.AnnotationInfoList
-
-interface CustomPreviewAnnotationLoader {
-
-    fun loadCustomPreviewAnnotation(): Map<String, AnnotationInfoList>
-}
