--- core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/classloader/classpath/previewfinder/overriden/annotationloader/PackageTreesCustomPreviewAnnotationLoader.kt
@@ -1,29 +1,0 @@
-package sergio.sastre.composable.preview.scanner.core.scanner.classloader.classpath.previewfinder.overriden.annotationloader
-
-import io.github.classgraph.AnnotationInfoList
-import io.github.classgraph.ClassGraph
-
-class PackageTreesCustomPreviewAnnotationLoader(
-    private val packageTreesOfCustomPreviews: List<String> = emptyList(),
-    private val annotationToScanClassName: String
-) : CustomPreviewAnnotationLoader {
-
-    private val scanResult
-        get() = ClassGraph()
-            .acceptPackages(*packageTreesOfCustomPreviews.toTypedArray())
-            .enableAnnotationInfo()
-            .scan()
-
-    override fun loadCustomPreviewAnnotation(): Map<String, AnnotationInfoList> =
-        when (packageTreesOfCustomPreviews.isEmpty()) {
-            true -> emptyMap()
-            false -> scanResult
-                .use { scanResult ->
-                    scanResult.allAnnotations
-                        .associate { clazz ->
-                            clazz.name to clazz.annotationInfo.filter { it.name == annotationToScanClassName }
-                        }.filterValues { it.isNotEmpty() }
-                }
-        }
-
-}
