+++ core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/config/classloader/ClassLoader.kt
@@ -1,0 +1,7 @@
+package sergio.sastre.composable.preview.scanner.core.scanner.config.classloader
+
+import io.github.classgraph.ClassInfo
+
+internal interface ClassLoader {
+    fun loadClass(classInfo: ClassInfo): Class<*>
+}
