+++ core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/config/classloader/classpath/Classpath.kt
@@ -1,0 +1,6 @@
+package sergio.sastre.composable.preview.scanner.core.scanner.config.classloader.classpath
+
+data class Classpath(
+    val packagePath: String,
+    val rootDir: String = "${System.getProperty("user.dir")}/build"
+)
