+++ core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanner/config/classloader/classpath/previewfinder/overriden/annotationloader/CustomPreviewAnnotationLoader.kt
@@ -1,0 +1,8 @@
+package sergio.sastre.composable.preview.scanner.core.scanner.config.classloader.classpath.previewfinder.overriden.annotationloader
+
+import io.github.classgraph.AnnotationInfoList
+
+interface CustomPreviewAnnotationLoader {
+
+    fun loadCustomPreviewAnnotation(): Map<String, AnnotationInfoList>
+}
