+++ core/src/main/java/sergio/sastre/composable/preview/scanner/core/scanresult/filter/ScanResultFilter.kt
@@ -1,9 +1,9 @@
 package sergio.sastre.composable.preview.scanner.core.scanresult.filter
 
-import io.github.classgraph.AnnotationInfoList
+
 import io.github.classgraph.ScanResult
 import sergio.sastre.composable.preview.scanner.core.preview.ComposablePreview
-import sergio.sastre.composable.preview.scanner.core.scanner.classloader.classpath.previewfinder.PreviewsFinder
+import sergio.sastre.composable.preview.scanner.core.scanner.config.classloader.classpath.previewfinder.PreviewsFinder
 
 /**
  * Filter the ComposablePreviews of a given ScanResult.
