+++ custompreviews/src/main/java/sergio/sastre/composable/preview/custompreviews/CrossModuleCustomPreview.kt
@@ -1,0 +1,9 @@
+package sergio.sastre.composable.preview.custompreviews
+
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewLightDark
+
+// 3 Previews
+@PreviewLightDark
+@Preview(group = "CrossModuleCustomPreview", fontScale = 1.1f)
+annotation class CrossModuleCustomPreview
