+++ settings.gradle.kts
@@ -26,3 +26,4 @@ include(":tests")
 include(":core")
 include(":android")
 include(":jvm")
+include(":custompreviews")
