+++ tests/src/main/java/sergio/sastre/composable/preview/scanner/multiplepreviewscrossmodule/Composables.kt
@@ -1,0 +1,11 @@
+package sergio.sastre.composable.preview.scanner.multiplepreviewscrossmodule
+
+import androidx.compose.runtime.Composable
+import sergio.sastre.composable.preview.custompreviews.CrossModuleCustomPreview
+import sergio.sastre.composable.preview.scanner.previewparameters.Example
+
+@CrossModuleCustomPreview // 3 Previews
+@Composable
+fun ExampleMultiplePreviewWithParams(){
+    Example("Example")
+}
